refactor: share table printing function (dry)
